# Conflicts:
#	src/main/java/com/example/PCOnlineShop/config/auth/SecurityConfig.java

+++ src/main/java/com/example/PCOnlineShop/repository/order/OrderRepository.java
@@ -9,10 +9,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-
 import java.util.List;
 
-public interface OrderRepository extends JpaRepository<Order, Integer> {
+// ✅ SỬA LỖI: Thay JpaRepository<Order, Integer> thành JpaRepository<Order, Long>
+public interface OrderRepository extends JpaRepository<Order, Long> {
 
     // --- Phương thức cho Staff List ---
     @Query(value = "SELECT o FROM Order o JOIN FETCH o.account a WHERE a.role = :role",
